Refactor export implementation to Util class
