+++ src/main/java/seedu/address/logic/commands/ExportCommand.java
@@ -3,20 +3,15 @@ package seedu.address.logic.commands;
 import static java.util.Objects.requireNonNull;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ATTRIBUTE;
 
-import java.io.BufferedWriter;
-import java.io.FileWriter;
 import java.io.IOException;
 import java.util.Arrays;
-import java.util.EnumMap;
 import java.util.List;
-import java.util.Objects;
-import java.util.function.Function;
-import java.util.stream.Collectors;
 
 import seedu.address.commons.util.ToStringBuilder;
 import seedu.address.logic.commands.exceptions.CommandException;
+import seedu.address.model.Attribute;
 import seedu.address.model.Model;
-import seedu.address.model.person.Person;
+import seedu.address.model.util.ExportDataUtil;
 
 /**
  * Edits the details of an existing person in the address book.
@@ -37,20 +32,6 @@ public class ExportCommand extends Command {
     public static final String MESSAGE_EXPORT_FAILURE = "Export failed";
     public static final String MESSAGE_EMPTY_LIST = "List is empty, nothing to export";
 
-    // Should we change this into a parameter of the export command?
-    public static final String DEFAULT_FILE_NAME = "export.csv";
-
-    // We have to use Object here since none of name, email, etc. have a common interface
-    private static final EnumMap<Attribute, Function<Person, Object>> PERSON_METHOD_MAP =
-            new EnumMap<>(Attribute.class);
-    static {
-        PERSON_METHOD_MAP.put(Attribute.NAME, Person::getName);
-        PERSON_METHOD_MAP.put(Attribute.PHONE, Person::getPhone);
-        PERSON_METHOD_MAP.put(Attribute.EMAIL, Person::getEmail);
-        PERSON_METHOD_MAP.put(Attribute.ADDRESS, Person::getAddress);
-        // TODO PERSON_METHOD_MAP.put(Attribute.COMMISSION, Person::getCommission);
-    }
-
     // This cannot be Attribute[] due to the given parser
     private final List<Attribute> attributes;
 
@@ -62,7 +43,7 @@ public class ExportCommand extends Command {
      */
     public ExportCommand(List<Attribute> attributes) {
         if (attributes == null || attributes.isEmpty()) {
-            this.attributes = Arrays.stream(Attribute.values()).toList();
+            this.attributes = Arrays.asList(Attribute.values());
         } else {
             this.attributes = attributes;
         }
@@ -71,32 +52,14 @@ public class ExportCommand extends Command {
     @Override
     public CommandResult execute(Model model) throws CommandException {
         requireNonNull(model);
-        List<Person> lastShownList = model.getFilteredPersonList();
-
-        if (lastShownList.isEmpty()) {
+        if (model.getFilteredPersonList().isEmpty()) {
             throw new CommandException(MESSAGE_EMPTY_LIST);
         }
-
-        // TODO discuss if we separate this into a component
-        try (BufferedWriter writer = new BufferedWriter(new FileWriter(DEFAULT_FILE_NAME))) {
-            // CSV header
-            writer.write(attributes.stream()
-                    .map(Objects::toString)
-                    .collect(Collectors.joining(",")));
-            writer.newLine();
-            // CSV body
-            for (Person person : lastShownList) {
-                writer.write(attributes.stream()
-                        .map(attribute -> PERSON_METHOD_MAP.get(attribute)
-                                .apply(person)
-                                .toString())
-                        .collect(Collectors.joining(",")));
-                writer.newLine();
-            }
+        try {
+            ExportDataUtil.exportAsCsv(model, this.attributes);
         } catch (IOException e) {
             throw new CommandException(MESSAGE_EXPORT_FAILURE, e);
         }
-
         return new CommandResult(MESSAGE_EXPORT_SUCCESS);
     }
 
@@ -105,17 +68,5 @@ public class ExportCommand extends Command {
         return new ToStringBuilder(this)
                 .add("attributes", attributes)
                 .toString();
-    }
-
-    /**
-     * Stores the details to edit the person with. Each non-empty field value will replace the
-     * corresponding field value of the person.
-     */
-    public enum Attribute {
-        NAME,
-        PHONE,
-        EMAIL,
-        ADDRESS,
-        // TODO COMMISSION,
     }
 }
