+++ src/main/java/seedu/address/logic/parser/ExportCommandParser.java
@@ -1,19 +1,19 @@
 package seedu.address.logic.parser;
 
 import static java.util.Objects.requireNonNull;
+import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 import static seedu.address.logic.parser.CliSyntax.PREFIX_ATTRIBUTE;
 
 import java.util.List;
 
 import seedu.address.logic.commands.ExportCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
+import seedu.address.model.Attribute;
 
 /**
  * Parses input arguments and creates a new ExportCommand object
  */
 public class ExportCommandParser implements Parser<ExportCommand> {
-    public static final String MESSAGE_INVALID_ATTRIBUTE = "Invalid attribute";
-
     /**
      * Parses the given {@code String} of arguments in the context of the ExportCommand
      * and returns an ExportCommand object for execution.
@@ -24,11 +24,11 @@ public class ExportCommandParser implements Parser<ExportCommand> {
         ArgumentMultimap argMultimap =
                 ArgumentTokenizer.tokenize(args, PREFIX_ATTRIBUTE);
         try {
-            List<ExportCommand.Attribute> attributes = argMultimap.getAllValues(PREFIX_ATTRIBUTE).stream()
-                    .map(ExportCommand.Attribute::valueOf).toList();
+            List<Attribute> attributes = argMultimap.getAllValues(PREFIX_ATTRIBUTE).stream()
+                    .map(Attribute::valueOf).toList();
             return new ExportCommand(attributes);
         } catch (IllegalArgumentException e) {
-            throw new ParseException(MESSAGE_INVALID_ATTRIBUTE, e);
+            throw new ParseException(String.format(MESSAGE_INVALID_COMMAND_FORMAT, ExportCommand.MESSAGE_USAGE));
         }
     }
 }
