+++ src/main/java/seedu/address/model/Attribute.java
@@ -1,0 +1,13 @@
+package seedu.address.model;
+
+/**
+ * Stores the details to edit the person with. Each non-empty field value will replace the
+ * corresponding field value of the person.
+ */
+public enum Attribute {
+    NAME,
+    PHONE,
+    EMAIL,
+    ADDRESS,
+    COMMISSION,
+}
